src/requirements.txt
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9.2
dnspython==1.16.0
eventlet==0.25.1
Flask==1.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SocketIO==3.1.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
json-rpc==1.13.0
jsonrpclib==0.1.7
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
mkdocs==1.1
mkdocs-material==5.0.2
monotonic==1.5
nexmo==2.4.0
nltk==3.4.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymdown-extensions==7.0
python-engineio==3.11.2
python-prctl==1.7
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.22.0
RPi.GPIO==0.7.0
six==1.14.0
SQLAlchemy==1.3.13
tornado==6.0.4
transitions==0.7.2
urllib3==1.25.8
websocket==0.2.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.1
WTForms==2.2.1
